# config/openapi_schema_snapshot.json
{
...
  "openapi": "3.1.0",
...
  "paths": {
...
    "/api/v1/analyze/process/rules/audit": {
      "get": {
        "deprecated": false,
        "operationId": "process_rules_audit_v1",
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
  },
...
}